removed redundant color enum
